requirements.txt
# Django & REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0

# PostgreSQL driver
psycopg2-binary>=2.9,<3.0

# Background task processing
celery>=5.3,<6.0
redis>=5.0,<6.0

# Environment variables
python-decouple>=3.8,<4.0

# Excel file parsing
openpyxl>=3.1,<4.0

# Date utilities
python-dateutil>=2.8,<3.0

# Production WSGI server
gunicorn>=21.2,<23.0
